Maintenance: migrate gem to Bundler-like structure
