+++ .gitignore
@@ -1,4 +1,11 @@
-*.gem
-.eprj
-pkg
-coverage
+/.bundle/
+/.yardoc
+/_yardoc/
+/coverage/
+/doc/
+/pkg/
+/spec/reports/
+/tmp/
+
+# rspec failure tracking
+.rspec_status
