+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+## [Unreleased]
+
+## [0.1.0] - 2023-02-25
+
+- Initial release
