+++ Rakefile
@@ -1,37 +1,12 @@
-require 'rake'
-require 'spec/rake/spectask'
+# frozen_string_literal: true
 
-desc "Run all examples with RCov"
-Spec::Rake::SpecTask.new('rcov') do |t|
-  t.spec_files = FileList['spec/*.rb']
-  t.rcov = true
-  t.rcov_opts = ['--exclude', 'spec']
-end
+require "bundler/gem_tasks"
+require "rspec/core/rake_task"
 
-begin
-  require 'bones'
-  Bones.setup
-rescue LoadError
-  load 'tasks/setup.rb'
-end
+RSpec::Core::RakeTask.new(:spec)
 
-ensure_in_path 'lib'
-require 'bones'
+require "rubocop/rake_task"
 
-task :default => 'spec:run'
+RuboCop::RakeTask.new
 
-PROJ.name = 'slither'
-PROJ.authors = 'Ryan Wood'
-PROJ.email = 'ryan.wood@gmail.com'
-PROJ.url = 'http://github.com/ryanwood/slither'
-PROJ.version = '0.99.3'
-PROJ.exclude = %w(\.git .gitignore ^tasks \.eprj ^pkg)
-PROJ.readme_file = 'README.rdoc'
-
-#PROJ.rubyforge.name = 'codeforpeople'
-
-PROJ.rdoc.exclude << '^data'
-PROJ.notes.exclude = %w(^README\.rdoc$ ^data ^pkg)
-
-# PROJ.svn.path = 'bones'
-# PROJ.spec.opts << '--color'
+task default: %i[spec rubocop]
