+++ lib/slither/column.rb
@@ -1,6 +1,6 @@
 require 'date'
 
-class Slither
+module Slither
   class ParserError < RuntimeError; end
 
   class Column
@@ -75,7 +75,7 @@ class Slither
       def inspect
         "#<#{self.class} #{instance_variables.map{|iv| "#{iv}=>#{instance_variable_get(iv)}"}.join(', ')}>"
       end
-      
+
       def to_s(value)
         result = case @type
           when :date
