+++ lib/slither/section.rb
@@ -1,10 +1,10 @@
-class Slither
+module Slither
   class Section
     attr_accessor :definition, :optional
     attr_reader :name, :columns, :options, :length
-    
+
     RESERVED_NAMES = [:spacer]
-    
+
     def initialize(name, options = {})
       @name = name
       @options = options
@@ -13,7 +13,7 @@ class Slither
       @optional = options[:optional] || false
       @length = 0
     end
-    
+
     def column(name, length, options = {})
       raise(Slither::DuplicateColumnNameError, "You have already defined a column named '#{name}'.") if @columns.map do |c|
         RESERVED_NAMES.include?(c.name) ? nil : c.name
@@ -23,15 +23,15 @@ class Slither
       @length += length
       col
     end
-    
+
     def spacer(length)
       column(:spacer, length)
     end
-    
+
     def trap(&block)
       @trap = block
     end
-    
+
     def template(name)
       template = @definition.templates[name]
       raise ArgumentError, "Template #{name} not found as a known template." unless template
@@ -40,18 +40,18 @@ class Slither
       # Section options should trump template options
       @options = template.options.merge(@options)
     end
-    
+
     def format(data)
       # raise( ColumnMismatchError,
       #   "The '#{@name}' section has #{@columns.size} column(s) defined, but there are #{data.size} column(s) provided in the data."
       # ) unless @columns.size == data.size
-      row = ''      
+      row = ''
       @columns.each do |column|
         row += column.format(data[column.name])
       end
       row
     end
-    
+
     def parse(line)
       line_data = line.unpack(unpacker)
       row = {}
@@ -60,7 +60,7 @@ class Slither
       end
       row
     end
-    
+
     def parse_when_problem(line)
       line_data = line.unpack(@columns.map { |c| "a#{c.length}" }.join(''))
       row = ''
@@ -69,20 +69,20 @@ class Slither
       end
       row
     end
-    
+
     def match(raw_line)
       raw_line.nil? ? false : @trap.call(raw_line)
     end
-    
+
     def method_missing(method, *args)
       column(method, *args)
     end
-  
+
     private
-      
+
       def unpacker
         @columns.map { |c| c.unpacker }.join('')
       end
 
-  end  
+  end
 end
