+++ lib/slither/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Slither
+  VERSION = "0.99.5"
+end
