+++ spec/slither_spec.rb
@@ -1,22 +1,20 @@
-require File.join(File.dirname(__FILE__), 'spec_helper')
-
-describe Slither do
-  
+RSpec.describe Slither do
   before(:each) do
     @name = :doc
     @options = { :align => :left }
   end
-  
+
   describe "when defining a format" do
-    before(:each) do
-      @definition = mock('definition')
+    let!(:definition) do
+      double("definition")
     end
-  
+
     it "should create a new definition using the specified name and options" do
-      Slither.should_receive(:define).with(@name, @options).and_return(@definition)
+      expect(described_class).to receive(:define).with(@name, @options).and_return(definition)
+
       Slither.define(@name , @options)
     end
-    
+
     it "should pass the definition to the block" do
       yielded = nil
       Slither.define(@name) do |y|
@@ -24,7 +22,7 @@ describe Slither do
       end
       yielded.should be_a( Slither::Definition )
     end
-    
+
     it "should add to the internal definition count" do
       Slither.definitions.clear
       Slither.should have(0).definitions
@@ -32,12 +30,12 @@ describe Slither do
       Slither.should have(1).definitions
     end
   end
-  
-  describe "when creating file from data" do 
+
+  describe "when creating file from data" do
     it "should raise an error if the definition name is not found" do
-      lambda { Slither.generate(:not_there, {}) }.should raise_error(ArgumentError)      
+      lambda { Slither.generate(:not_there, {}) }.should raise_error(ArgumentError)
     end
-    
+
     it "should output a string" do
       definition = mock('definition')
       generator = mock('generator')
@@ -46,7 +44,7 @@ describe Slither do
       Slither::Generator.should_receive(:new).with(definition).and_return(generator)
       Slither.generate(:test, {})
     end
-    
+
     it "should output a file" do
   	  file = mock('file')
   	  text = mock('string')
@@ -54,35 +52,35 @@ describe Slither do
   	  File.should_receive(:open).with('file.txt', 'w').and_yield(file)
   	  Slither.should_receive(:generate).with(:test, {}).and_return(text)
       Slither.write('file.txt', :test, {})
-  	end       
+  	end
   end
-	
+
   describe "when parsing a file" do
     before(:each) do
       @file_name = 'file.txt'
     end
-    
+
     it "should check the file exists" do
       lambda { Slither.parse(@file_name, :test, {}) }.should raise_error(ArgumentError)
     end
-    
+
     it "should raise an error if the definition name is not found" do
       Slither.definitions.clear
       File.stub!(:exists? => true)
-      lambda { Slither.parse(@file_name, :test, {}) }.should raise_error(ArgumentError)      
+      lambda { Slither.parse(@file_name, :test, {}) }.should raise_error(ArgumentError)
     end
-    
+
     it "should create a parser and call parse" do
       File.stub!(:exists? => true)
       file_io = mock("IO")
       parser = mock("parser")
       definition = Slither::Definition.new :by_bytes => false
-      
+
       File.should_receive(:open).and_return(file_io)
       Slither.should_receive(:definition).with(:test).and_return(definition)
-      Slither::Parser.should_receive(:new).with(definition, file_io).and_return(parser)    
+      Slither::Parser.should_receive(:new).with(definition, file_io).and_return(parser)
       parser.should_receive(:parse).and_return("parse result")
-      
+
       Slither.parse(@file_name, :test).should eq("parse result")
     end
   end
