+++ spec/spec_helper.rb
@@ -1,3 +1,15 @@
-require 'rubygems'
-require File.expand_path(File.join(File.dirname(__FILE__), '..', 'lib', 'slither'))
- 
+# frozen_string_literal: true
+
+require "slither"
+
+RSpec.configure do |config|
+  # Enable flags like --only-failures and --next-failure
+  config.example_status_persistence_file_path = ".rspec_status"
+
+  # Disable RSpec exposing methods globally on `Module` and `main`
+  config.disable_monkey_patching!
+
+  config.expect_with :rspec do |c|
+    c.syntax = :expect
+  end
+end
